@@ -1,6 +1,6 @@
 {
   "name": "admin-template",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "description": "",
   "main": "index.js",
   "private": true,
@@ -15,10 +15,11 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@vue/compiler-sfc": "^3.0.11",
     "axios": "^0.18.1",
     "dotenv-webpack": "^1.7.0",
     "echarts": "^4.9.0",
-    "element-ui": "2.13.2",
+    "element-plus": "^1.0.2-beta.40",
     "fingerprintjs2": "^1.8.6",
     "js-base64": "^2.5.1",
     "lockr": "^0.8.5",
@@ -26,26 +27,27 @@
     "mavon-editor": "^2.6.17",
     "moment": "^2.24.0",
     "nprogress": "^0.2.0",
-    "qrcode.vue": "^1.7.0",
+    "qrcode.vue": "3.2.*",
     "video.js": "^7.6.0",
     "videojs-brand": "0.0.4",
     "videojs-contextmenu-ui": "^5.1.1",
     "videojs-playlist": "^4.3.1",
     "videojs-playlist-ui": "^3.5.2",
-    "vue": "^2.6.10",
-    "vue-i18n": "^8.12.0",
-    "vue-json-pretty": "^1.6.0",
-    "vue-router": "^3.0.7",
-    "vuex": "^3.1.1"
+    "vue": "^3.0.11",
+    "vue-i18n": "^9.1.*",
+    "vue-loader": "^16.2.0",
+    "vue-router": "^4.0.6",
+    "vue-style-loader": "^4.1.3",
+    "vuex": "^4.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
-    "@vue/test-utils": "^1.0.0-beta.29",
+    "@vue/test-utils": "^2.0.0-rc.6",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.2",
     "babel-helper-vue-jsx-merge-props": "^2.0.3",
-    "babel-jest": "^23.6.0",
+    "babel-jest": "^25.5.1",
     "babel-loader": "^8.0.6",
     "babel-plugin-component": "^1.1.1",
     "babel-plugin-syntax-jsx": "^6.18.0",
@@ -65,7 +67,7 @@
     "eslint-plugin-vue": "^5.2.3",
     "file-loader": "^3.0.1",
     "html-webpack-plugin": "^3.2.0",
-    "jest": "^23.6.0",
+    "jest": "^24.9.0",
     "jest-serializer-vue": "^2.0.2",
     "mini-css-extract-plugin": "^0.5.0",
     "style-loader": "^0.23.1",
@@ -73,10 +75,8 @@
     "stylus-loader": "^3.0.2",
     "uglify-es": "^3.3.9",
     "url-loader": "^1.1.2",
-    "vue-jest": "^3.0.4",
-    "vue-loader": "^15.7.1",
+    "vue-jest": "^5.0.0-alpha.7",
     "vue-server-renderer": "^2.6.10",
-    "vue-style-loader": "^4.1.2",
     "vue-template-compiler": "^2.6.10",
     "webpack": "^4.37.0",
     "webpack-cli": "^3.3.6",
