@@ -62,8 +62,8 @@
 </template>
 
 <script>
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
 
 	export default {
 		components: {
