@@ -251,8 +251,8 @@
 
 <script>
 	import api from '../../api/albumApi'
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
 
 	export default {
         props: {
