@@ -380,16 +380,13 @@
 
 <script>
 	import api from '../../api/bannersApi'
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
 	import ImgUploader from '@/components/img-uploader.vue'
 
 	export default {
 		components: {
-			ElForm,
-			ElButtonGroup,
-			ElButton,
             ImgUploader
 		},
 		data() {
