@@ -413,17 +413,14 @@
 <script>
 	import fileApi from '../../api/fileApi'
 	import articleApi from '../../api/articleApi'
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
 	import ImgUploader from '@/components/img-uploader.vue'
 
 
 	export default {
 		components: {
-			ElForm,
-			ElButtonGroup,
-			ElButton,
             ImgUploader
 		},
 		data() {
