@@ -5,26 +5,32 @@
         -moz-box-sizing: border-box;
         box-sizing: border-box;
     }
+
     *:before, *:after {
         -webkit-box-sizing: border-box;
         -moz-box-sizing: border-box;
         box-sizing: border-box;
     }
-    html{
+
+    html {
         height: 100%;
     }
-    body{
+
+    body {
         height: 100%;
         font-size: 12px;
     }
-    #app{
+
+    #app {
         height: 100%;
     }
-    .wrapper{
+
+    .wrapper {
         position: relative;
         height: auto;
         min-height: 100%;
     }
+
     .content-wrapper {
         position: absolute;
         top: 0;
@@ -41,68 +47,82 @@
         -moz-transition: all 0.2s ease;
         -webkit-transition: all 0.2s ease;
     }
-    .el-loading-mask{
-        z-index: 999!important;
+
+    .el-loading-mask {
+        z-index: 999 !important;
     }
-    .sidebar-collapse .content-wrapper{
+
+    .sidebar-collapse .content-wrapper {
         margin-left: 50px !important;
         z-index: 840;
     }
-    .content-wrapper .main-content{
+
+    .content-wrapper .main-content {
         margin: 10px 10px 0;
         padding: 0;
         min-height: calc(100vh - 101px);
         padding-bottom: 50px;
     }
-    .content-wrapper .main-content.iframe{
+
+    .content-wrapper .main-content.iframe {
         height: calc(100vh - 101px);
     }
 
-    [class^="icon-"], [class*=" icon-"]{
+    [class^="icon-"], [class*=" icon-"] {
         font-family: FontAwesome;
         -webkit-font-smoothing: antialiased;
         -moz-osx-font-smoothing: grayscale;
     }
+
     .main-topbar .topbar-clearfix:before, .main-topbar .topbar-clearfix:after {
         display: table;
         content: " ";
     }
-    .main-topbar{
+
+    .main-topbar {
         position: fixed;
         width: 100%;
         height: 50px;
         background: #373d41;
-      z-index: 50;
+        z-index: 50;
     }
-    .main-topbar .topbar-wrap{
+
+    .main-topbar .topbar-wrap {
         height: 100%;
     }
-    .main-topbar .topbar-head{
+
+    .main-topbar .topbar-head {
         background: #2a2f32;
         height: 50px;
         position: relative;
         z-index: 3;
     }
+
     .main-topbar .topbar-btn {
         color: #fff;
         font-size: 14px;
         line-height: 50px;
     }
+
     .main-topbar .topbar-logo span {
         line-height: 50px;
     }
+
     .main-topbar .topbar-left {
         float: left;
     }
+
     .icon-logo {
         font-size: 24px;
         line-height: 50px;
         display: block;
     }
+
     .main-topbar .topbar-logo {
         background: #373d41;
     }
-    .main-topbar .topbar-logo{
+
+    .main-topbar .topbar-logo {
         display: block;
         width: 50px;
         color: #FFF;
@@ -110,34 +130,41 @@
         line-height: 50px;
         margin-right: 1px;
     }
+
     .main-topbar .topbar-btn:hover, .main-topbar .topbar-btn.topbar-btn-dark {
         background: #2a2f32;
     }
+
     .main-topbar .topbar-info .topbar-btn:hover, .main-topbar .topbar-info .topbar-btn.topbar-btn-dark {
         background: #2a2f32;
     }
+
     .main-topbar .topbar-home-link {
         margin-right: 1px;
         background: #373d41;
         width: 129px;
         text-align: center;
     }
+
     .main-topbar .topbar-product {
         position: relative;
         z-index: 2;
         background: #373d41;
     }
+
     .main-topbar .topbar-product-btn {
         padding: 0 20px;
         display: inline-block;
         height: 50px;
         cursor: pointer;
     }
+
     .main-topbar span, .main-topbar a, .main-topbar p, .main-topbar li {
         text-rendering: optimizeLegibility;
         -webkit-font-smoothing: antialiased;
         letter-spacing: 0.02em;
     }
+
     .main-topbar .icon-arrow-down {
         display: inline-block;
         width: 18px;
@@ -149,9 +176,11 @@
         -moz-transition: transform 0.2s, vertical-align 0.2s;
         -webkit-transition: transform 0.2s, vertical-align 0.2s;
     }
+
     .icon-arrow-down:before {
         content: "\f0d7";
     }
+
     .main-topbar .topbar-info {
         background: #2a2f32;
         position: absolute;
@@ -159,20 +188,25 @@
         top: 0;
         right: 0;
     }
+
     .main-topbar .topbar-wrap, .main-topbar .topbar-logo, .main-topbar .topbar-home, .main-topbar .topbar-home-link, .main-topbar .topbar-product, .main-topbar .topbar-info {
         height: 100%;
     }
+
     .main-topbar .topbar-notice {
         position: relative;
         font-size: 12px;
     }
+
     .main-topbar .topbar-info-item {
         display: inline-block;
         position: relative;
     }
+
     .main-topbar .topbar-notice .topbar-btn {
         padding: 0 16px;
     }
+
     .main-topbar .topbar-info .topbar-btn {
         padding: 0 18px;
         height: 50px;
@@ -182,16 +216,20 @@
         background: #373d41;
         border-right: 1px solid #2a2f32;
     }
+
     .main-topbar .topbar-info-dropdown .topbar-btn {
         position: relative;
     }
+
     .main-topbar .topbar-btn-notice .topbar-btn-notice-icon {
         line-height: 50px;
         vertical-align: text-bottom;
     }
+
     .icon-bell {
         font-size: 20px;
     }
+
     .main-topbar .topbar-btn-notice .topbar-btn-notice-num {
         font-size: 12px;
         color: #fff;
@@ -204,14 +242,17 @@
         vertical-align: top;
         text-align: center;
     }
+
     .main-topbar .topbar-info-dropdown .topbar-info-dropdown-toggle {
         color: #fff;
     }
+
     .main-topbar span, .main-topbar a, .main-topbar p, .main-topbar li {
         text-rendering: optimizeLegibility;
         -webkit-font-smoothing: antialiased;
         letter-spacing: 0.02em;
     }
+
     .main-topbar .topbar-info-dropdown-memu {
         padding: 0;
         position: absolute;
@@ -234,10 +275,12 @@
         transition: opacity 0.15s, visibility 0s 0.15s;
         opacity: 0;
     }
-    .main-topbar .topbar-info-dropdown-memu.show{
+
+    .main-topbar .topbar-info-dropdown-memu.show {
         visibility: visible;
         opacity: 1;
     }
+
     .main-topbar .topbar-info-dropdown-memu li a {
         display: block;
         padding: 0 20px;
@@ -250,31 +293,37 @@
         -webkit-transition: all 0.15s;
         transition: all 0.15s;
     }
+
     .main-topbar .topbar-info .topbar-info-btn a {
         height: 40px;
         line-height: 39px;
         padding: 0 18px;
     }
+
     .main-topbar .topbar-info-dropdown-memu li a:hover, .main-topbar .topbar-info-dropdown-memu li a:focus {
         background-color: #f5f5f5;
     }
+
     .main-topbar .topbar-info-dropdown-memu-list {
         padding: 10px 0;
         margin: 0px;
     }
+
     .main-body {
         position: absolute;
         width: 100%;
         top: 50px;
         bottom: 0px;
         background-color: #000;
-      z-index: 40;
+        z-index: 40;
     }
+
     .main-sidebar span, .main-sidebar a, .main-sidebar p, .main-sidebar li {
         text-rendering: optimizeLegibility;
         -webkit-font-smoothing: antialiased;
         letter-spacing: 0.02em;
     }
+
     .main-sidebar-full .main-sidebar, .main-sidebar.sidebar-full {
         width: 180px;
         display: block;
@@ -287,24 +336,28 @@
         overflow-x: hidden;
         overflow-y: hidden;
     }
+
     .main-sidebar {
         width: 0;
         position: fixed;
         top: 50px;
         bottom: 0;
         background-color: #333744;
-      z-index: 2;
+        z-index: 2;
     }
+
     .main-sidebar .sidebar-trans {
         -o-transition: all 0.12s ease, 0.12s ease;
         -ms-transition: all 0.12s ease, 0.12s ease;
         -moz-transition: all 0.12s ease, 0.12s ease;
         -webkit-transition: all 0.12s ease, 0.12s ease;
     }
+
     .main-sidebar .sidebar-inner {
         position: relative;
         height: 100%;
     }
+
     .main-sidebar .sidebar-fold {
         height: 30px;
         width: 100%;
@@ -318,27 +371,33 @@
         -webkit-user-select: none;
         -moz-user-select: none;
     }
-    .sidebar-fold > span{
+
+    .sidebar-fold > span {
         display: inline-block;
     }
+
     .icon-unfold {
         display: block;
         text-align: center;
     }
-    .sidebar-fold .icon-fold{
+
+    .sidebar-fold .icon-fold {
         transform: rotate(90deg);
     }
+
     .main-sidebar .sidebar-nav-wrap {
         width: 180px;
         height: calc(100% - 30px);
         overflow-x: hidden;
         overflow-y: auto;
     }
+
     .main-sidebar .sidebar-nav {
         width: 100%;
         overflow-x: hidden;
         overflow-y: auto;
     }
+
     .main-sidebar .sidebar-nav li a {
         display: block;
         width: 180px;
@@ -346,9 +405,11 @@
         line-height: 40px;
         overflow: hidden;
     }
+
     .main-sidebar .sidebar-nav li a:hover {
         background: #4A5064;
     }
+
     .main-sidebar .sidebar-nav .sidebar-title {
         height: 40px;
         background: #42485B;
@@ -360,13 +421,16 @@
         -moz-user-select: none;
         overflow: hidden;
     }
+
     .main-sidebar .sidebar-nav .sidebar-title:hover {
         background: #00C1DE;
     }
+
     .main-sidebar .sidebar-nav-active .sidebar-title-icon {
         vertical-align: text-top;
         transform: rotate(90deg);
     }
+
     .main-sidebar .sidebar-nav .sidebar-title-icon {
         width: 45px;
         text-align: center;
@@ -381,9 +445,11 @@
         -moz-transition: -moz-transform 0.12s;
         -webkit-transition: -webkit-transform 0.12s;
     }
+
     .icon-arrow-right:before {
         content: "\f0da";
     }
+
     .main-sidebar .sidebar-nav ul {
         width: 200px;
         margin: 0;
@@ -392,12 +458,15 @@
         overflow-y: scroll;
         overflow-x: hidden;
     }
+
     .main-sidebar .sidebar-nav .nav-item {
         position: relative;
     }
+
     .main-sidebar-full .main-sidebar .sidebar-nav .nav-icon, .main-sidebar.sidebar-full .sidebar-nav .nav-icon {
         width: 50px;
     }
+
     .main-sidebar .sidebar-nav .nav-icon {
         width: 50px;
         text-align: center;
@@ -405,6 +474,7 @@
         float: left;
         color: #aeb9c2;
     }
+
     .main-sidebar .sidebar-nav .nav-title {
         float: left;
         overflow: hidden;
@@ -414,27 +484,34 @@
         display: block;
         width: 130px;
     }
+
     .main-sidebar-mini .main-sidebar, .main-sidebar.sidebar-mini {
         width: 50px;
         display: block;
     }
+
     .sidebar-content .sidebar-nav li.active a .nav-icon {
         color: #fff;
     }
+
     .main-sidebar-mini .sidebar-content .sidebar-nav ul li a, .main-sidebar-mini .sidebar-content .sidebar-nav ul li a {
         width: 50px;
     }
+
     .sidebar-content .sidebar-nav li.active a {
         background: #00C1DE;
         transition: none;
         -webkit-transition: none;
     }
+
     .main-sidebar-mini .main-sidebar .sidebar-title .sidebar-title-text, .main-sidebar.sidebar-mini .sidebar-title .sidebar-title-text {
         display: none;
     }
+
     .main-sidebar-mini .main-product {
         left: 50px;
     }
+
     .main-product {
         width: auto;
         position: absolute;
@@ -445,12 +522,15 @@
         overflow: hidden;
         background: #FFF;
     }
+
     .main-sidebar-full .main-product {
         left: 180px;
     }
+
     .main-product-col-1 .main-product-navbar-bg, .main-product-col-1 .main-product-navbar {
         width: 180px;
     }
+
     .main-product-navbar {
         width: 0;
         float: left;
@@ -465,6 +545,7 @@
         -moz-transition: all 0.2s ease;
         -webkit-transition: all 0.2s ease;
     }
+
     .main-product-navbar .product-nav-stage {
         width: 180px;
         overflow: hidden;
@@ -473,25 +554,30 @@
         bottom: 0;
         right: 0;
     }
+
     .main-product-navbar .product-nav-stage .product-nav-scene {
         width: 180px;
         position: absolute;
         top: 0;
         bottom: 0;
-        -webkit-transition: position,.2s,linear;
-        -moz-transition: position,.2s,linear;
+        -webkit-transition: position, .2s, linear;
+        -moz-transition: position, .2s, linear;
     }
+
     .main-product-navbar .product-nav-stage-main .product-nav-main-scene {
         left: 0;
     }
+
     .main-product-navbar .product-nav-main-scene .product-nav-title {
         font-weight: bold;
         text-indent: 20px;
     }
+
     .main-product-navbar .product-nav-main-scene .product-nav-title {
         font-weight: bold;
         text-indent: 20px;
     }
+
     .main-product-navbar .product-nav-scene .product-nav-title {
         width: 180px;
         height: 70px;
@@ -501,6 +587,7 @@
         text-overflow: ellipsis;
         white-space: nowrap;
     }
+
     .main-product-navbar .product-nav-list {
         position: absolute;
         top: 70px;
@@ -510,14 +597,17 @@
         overflow-y: auto;
         overflow-x: hidden;
     }
+
     .main-product-navbar .product-nav-list ul {
         list-style: none;
         padding: 0;
         margin: 0;
     }
+
     .main-product-navbar .product-nav-list li.active a {
         background-color: #FFF;
     }
+
     .main-product-navbar .product-nav-list li a {
         width: 180px;
         height: 40px;
@@ -525,9 +615,11 @@
         display: block;
         color: #333;
     }
+
     .main-product-navbar .product-nav-list li a:hover {
         background-color: #F4F6F8;
     }
+
     .main-product-navbar .product-nav-list .nav-icon {
         width: 30px;
         height: 40px;
@@ -536,6 +628,7 @@
         font-size: 16px;
         color: #333;
     }
+
     .main-product-navbar .product-nav-list .nav-title {
         width: 138px;
         float: left;
@@ -543,12 +636,15 @@
         text-overflow: ellipsis;
         white-space: nowrap;
     }
+
     .main-product-navbar .product-nav-list ul ul li a {
         color: #666;
     }
+
     .main-product-col-1 .main-product-navbar-collapse {
         left: 160px;
     }
+
     .main-product-navbar-collapse {
         position: absolute;
         left: 0;
@@ -560,15 +656,18 @@
         -moz-transition: all 0.2s ease;
         -webkit-transition: all 0.2s ease;
     }
-    .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse-inner{
+
+    .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse-inner {
         z-index: 2;
     }
+
     .main-product-navbar-collapse .product-navbar-collapse-inner {
         top: -50%;
         position: relative;
         overflow: hidden;
         z-index: 3;
     }
+
     .main-product-navbar-collapse .product-navbar-collapse-bg {
         width: 0;
         height: 50px;
@@ -578,20 +677,23 @@
         border-bottom: 9px solid transparent;
         border-left: 13px solid #D9DEE4;
         border-top: 9px solid transparent;
-        -o-transition: all 0.1s ease,0.1s ease;
-        -ms-transition: all 0.1s ease,0.1s ease;
-        -moz-transition: all 0.1s ease,0.1s ease;
-        -webkit-transition: all 0.1s ease,0.1s ease;
+        -o-transition: all 0.1s ease, 0.1s ease;
+        -ms-transition: all 0.1s ease, 0.1s ease;
+        -moz-transition: all 0.1s ease, 0.1s ease;
+        -webkit-transition: all 0.1s ease, 0.1s ease;
     }
+
     .main-product-navbar-collapse:hover .product-navbar-collapse-bg {
         border-bottom: 8px solid transparent;
         border-left: 20px solid #D9DEE4;
         border-top: 8px solid transparent;
     }
+
     .main-product-navbar-collapse:hover .product-navbar-collapse {
         left: 0;
         font-size: 16px;
     }
+
     .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse-bg {
         right: 0;
         left: auto;
@@ -600,70 +702,86 @@
         border-right: 13px solid #f7f7f7;
         border-top: 9px solid transparent;
     }
+
     .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse {
         right: -7px;
         left: auto;
     }
-    .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse>span {
+
+    .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse > span {
         color: #546478;
     }
+
     .main-product-col-1 .main-product-navbar-collapse .icon-collapse-left {
         display: inline;
     }
-    .main-product-navbar-collapse .product-navbar-collapse>span {
+
+    .main-product-navbar-collapse .product-navbar-collapse > span {
         font-size: 20px;
         line-height: 50px;
         vertical-align: text-top;
     }
+
     .icon-collapse-left:before {
         content: "\f100";
     }
+
     .icon-collapse-right:before {
         content: "\f101";
     }
-    .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse>span {
+
+    .main-product-col-1 .main-product-navbar-collapse .product-navbar-collapse > span {
         color: #546478;
     }
+
     .main-product-col-1 .main-product-navbar-collapse .icon-collapse-left {
         display: inline;
     }
+
     .main-product-col-1 .main-product-navbar-collapse .icon-collapse-right {
         display: inline;
     }
+
     .main-product-navbar-collapse .product-navbar-collapse {
         height: 50px;
         position: relative;
         left: -7px;
         text-align: center;
         cursor: pointer;
-        -o-transition: all 0.1s ease,0.1s ease;
-        -ms-transition: all 0.1s ease,0.1s ease;
-        -moz-transition: all 0.1s ease,0.1s ease;
-        -webkit-transition: all 0.1s ease,0.1s ease;
+        -o-transition: all 0.1s ease, 0.1s ease;
+        -ms-transition: all 0.1s ease, 0.1s ease;
+        -moz-transition: all 0.1s ease, 0.1s ease;
+        -webkit-transition: all 0.1s ease, 0.1s ease;
     }
+
     .main-product-col-1 .main-product-navbar-collapse:hover .product-navbar-collapse {
         right: 0;
         left: auto;
     }
+
     .main-product-col-1 .main-product-navbar-collapse:hover .product-navbar-collapse-bg {
         border-bottom: 8px solid transparent;
         border-left: none;
         border-right: 20px solid #f7f7f7;
         border-top: 8px solid transparent;
     }
+
     .fade {
         opacity: 0;
         -webkit-transition: opacity 0.15s linear;
         transition: opacity 0.15s linear;
     }
+
     .main-sidebar-tooltip.right {
         padding: 0 5px;
         margin-left: 3px;
     }
+
     .main-sidebar-tooltip.in {
         opacity: 0.9;
         filter: alpha(opacity=90);
     }
+
     .main-sidebar-tooltip {
         position: absolute;
         z-index: 1030;
@@ -674,6 +792,7 @@
         filter: alpha(opacity=0);
         visibility: visible;
     }
+
     .main-sidebar-tooltip.right .tooltip-arrow {
         top: 50%;
         left: 0;
@@ -681,6 +800,7 @@
         border-right-color: #425160;
         border-width: 5px 5px 5px 0;
     }
+
     .main-sidebar-tooltip .tooltip-arrow {
         position: absolute;
         width: 0;
@@ -688,6 +808,7 @@
         border-color: transparent;
         border-style: solid;
     }
+
     .main-sidebar-tooltip .tooltip-inner {
         max-width: 200px;
         padding: 12px 8px;
@@ -697,377 +818,378 @@
         border-radius: 0 0;
         background-color: #425160;
     }
+
     .el-badge__content.is-fixed {
-      position: absolute;
-      border-radius: 0px;
-      background: #373d41;
-      border: 0px;
-      top: 15px;
-      right: 2px;
+        position: absolute;
+        border-radius: 0px;
+        background: #373d41;
+        border: 0px;
+        top: 15px;
+        right: 2px;
     }
 
     .alert-small {
-      padding: 0 5px;
+        padding: 0 5px;
     }
 
     .alert-small .el-alert__title {
-      font-size: 12px;
-      line-height: 14px;
+        font-size: 12px;
+        line-height: 14px;
 
     }
 
 </style>
 <template>
-  <div
-    v-show="isLoading"
-    v-loading.fullscreen.lock="!isLoading"
-    class="wrapper skin-blue"
-  >
-    <!--头部信息-->
-    <div class="main-topbar">
-      <div class="topbar-wrap">
-        <div class="topbar-head topbar-left">
-          <a
-            class="topbar-btn topbar-logo topbar-left"
-            href="javascript:void(0);"
-            @click="jump2AdminIndex()"
-          >
-            <span class="by-icon icon-logo by-SaaSguanlikongzhitai" />
-          </a>
-          <a
-            href="javascript:void(0);"
-            class="topbar-home-link topbar-btn topbar-left"
-            @click="jump2AdminIndex()"
-          >
-            <span>{{ platformInfo.all }}</span>
-          </a>
-        </div>
+    <div
+            v-show="isLoading"
+            v-loading.fullscreen.lock="!isLoading"
+            class="wrapper skin-blue"
+    >
+        <!--头部信息-->
+        <div class="main-topbar">
+            <div class="topbar-wrap">
+                <div class="topbar-head topbar-left">
+                    <a
+                            class="topbar-btn topbar-logo topbar-left"
+                            href="javascript:void(0);"
+                            @click="jump2AdminIndex()"
+                    >
+                        <span class="by-icon icon-logo by-SaaSguanlikongzhitai"/>
+                    </a>
+                    <a
+                            href="javascript:void(0);"
+                            class="topbar-home-link topbar-btn topbar-left"
+                            @click="jump2AdminIndex()"
+                    >
+                        <span>{{ platformInfo.all }}</span>
+                    </a>
+                </div>
 
-        <div class="topbar-info topbar-right topbar-clearfix">
-          <div
-            style="display:none;"
-            class="dropdown topbar-notice topbar-left topbar-info-dropdown topbar-info-item"
-          >
-            <a
-              href="javascript:void(0);"
-              class="topbar-btn topbar-btn-notice topbar-hover-dark"
-            >
-              <span class="topbar-btn-notice-icon icon-bell by-icon by-xiaoxi" />
-              <span class="topbar-btn-notice-num">
+                <div class="topbar-info topbar-right topbar-clearfix">
+                    <div
+                            style="display:none;"
+                            class="dropdown topbar-notice topbar-left topbar-info-dropdown topbar-info-item"
+                    >
+                        <a
+                                href="javascript:void(0);"
+                                class="topbar-btn topbar-btn-notice topbar-hover-dark"
+                        >
+                            <span class="topbar-btn-notice-icon icon-bell by-icon by-xiaoxi"/>
+                            <span class="topbar-btn-notice-num">
                 0
               </span>
-            </a>
-          </div>
-          <div class="hidden topbar-product topbar-left topbar-info-item">
-            <div
-              class="topbar-btn topbar-product-btn"
-              @click="iframeRefresh"
-            >
-              <span><i class="by-icon by-loading" />刷新</span>
-            </div>
-          </div>
+                        </a>
+                    </div>
+                    <div class="hidden topbar-product topbar-left topbar-info-item">
+                        <div
+                                class="topbar-btn topbar-product-btn"
+                                @click="iframeRefresh"
+                        >
+                            <span><i class="by-icon by-loading"/>刷新</span>
+                        </div>
+                    </div>
 
-          <div class="topbar-product topbar-left topbar-info-item">
-            <div
-                    class="topbar-btn topbar-product-btn"
-                    @click="goMessage"
-            >
-              <el-badge :value="unreadMsgCnt" :max="99" class="item">
-                <i class="el-icon-bell" />
-              </el-badge>
+                    <div class="topbar-product topbar-left topbar-info-item">
+                        <div
+                                class="topbar-btn topbar-product-btn"
+                                @click="goMessage"
+                        >
+                            <el-badge :value="unreadMsgCnt" :max="99" class="item">
+                                <i class="el-icon-bell"/>
+                            </el-badge>
+                        </div>
+                    </div>
+                    <TopBarDropMenu
+                            :name="userInfo.mobile"
+                            :head="userInfo.head"
+                            iconClass=""
+                            :links="userDropMenus"
+                            @logout="logout"
+                    />
+                    <TopBarDropMenu
+                            name=""
+                            iconClass="by-icon by-duoyuyan"
+                            :links="languages"
+                            @changeLanguages="changeLanguages"
+                    />
+                    <div class="topbar-product topbar-left topbar-info-item">
+                        <div
+                                class="topbar-btn topbar-product-btn"
+                                @click="toggleFullScreen"
+                        >
+                            <i title="Fullscreen F11" class="by-icon by-webtubiaoku10"></i>
+                        </div>
+                    </div>
+                </div>
             </div>
-          </div>
-          <TopBarDropMenu
-            :name="userInfo.mobile"
-            :head="userInfo.head"
-            iconClass=""
-            :links="userDropMenus"
-            @logout="logout"
-          />
-            <TopBarDropMenu
-                    name=""
-                    iconClass="by-icon by-duoyuyan"
-                    :links="languages"
-                    @changeLanguages="changeLanguages"
-            />
-          <div class="topbar-product topbar-left topbar-info-item">
-            <div
-              class="topbar-btn topbar-product-btn"
-              @click="toggleFullScreen"
-            >
-              <i title="Fullscreen F11" class="by-icon by-webtubiaoku10"></i>
-            </div>
-          </div>
         </div>
-      </div>
-    </div>
-    <div
-      class="main-body"
-      :class="{ 'main-sidebar-mini': isMiniMenu, 'main-sidebar-full': !isMiniMenu}"
-    >
-      <!--左边导航-->
-      <div class="main-sidebar">
-        <div class="sidebar-content main-sidebar">
-          <div class="sidebar-inner">
-            <div
-              class="sidebar-fold"
-              @click="toggleMiniMenu"
-            >
-              <span
-                class="by-icon"
-                :class="isMiniMenu ? 'by-fold' : 'by-santiaogang'"
-              />
-            </div>
-            <div class="sidebar-nav-wrap">
-              <SideBarNav
-                v-for="(menu, index) in menuList"
-                :key="menu.Id"
-                :is-nav-active="activeNavIndex === index"
-                :nav-menu="menu"
-                :index="index"
-                :nav-max-height="navMaxHeight"
-                :active-index="activeNavMenuIndex"
-                @NavOnClick="SideBarNavOnClick"
-                @NavOnActive="SideBarNavOnActive"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-      <!--产品部分-->
-      <div
-        class="main-product"
-        :class="{ 'main-product-col-1': isShowSecondMenu }"
-      >
-        <SecondNavBar
-          v-if="secondMenuData"
-          :is-show-second-menu="isShowSecondMenu"
-          :menu="secondMenuData"
-          @isShowSecondMenu="toggleSecondMenu"
-        />
-        <!--中间内容-->
         <div
-          class="content-wrapper"
-          :style="{left: isShowSecondMenu ? 180 + 'px' : '0'}"
+                class="main-body"
+                :class="{ 'main-sidebar-mini': isMiniMenu, 'main-sidebar-full': !isMiniMenu}"
         >
-          <transition
-            name="custom-classes-transition"
-            enter-active-class="animated fadeIn"
-          >
-            <slot name="main-content" />
-          </transition>
+            <!--左边导航-->
+            <div class="main-sidebar">
+                <div class="sidebar-content main-sidebar">
+                    <div class="sidebar-inner">
+                        <div
+                                class="sidebar-fold"
+                                @click="toggleMiniMenu"
+                        >
+              <span
+                      class="by-icon"
+                      :class="isMiniMenu ? 'by-fold' : 'by-santiaogang'"
+              />
+                        </div>
+                        <div class="sidebar-nav-wrap">
+                            <SideBarNav
+                                    v-for="(menu, index) in menuList"
+                                    :key="menu.Id"
+                                    :is-nav-active="activeNavIndex === index"
+                                    :nav-menu="menu"
+                                    :index="index"
+                                    :nav-max-height="navMaxHeight"
+                                    :active-index="activeNavMenuIndex"
+                                    @NavOnClick="SideBarNavOnClick"
+                                    @NavOnActive="SideBarNavOnActive"
+                            />
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <!--产品部分-->
+            <div
+                    class="main-product"
+                    :class="{ 'main-product-col-1': isShowSecondMenu }"
+            >
+                <SecondNavBar
+                        v-if="secondMenuData"
+                        :is-show-second-menu="isShowSecondMenu"
+                        :menu="secondMenuData"
+                        @isShowSecondMenu="toggleSecondMenu"
+                />
+                <!--中间内容-->
+                <div
+                        class="content-wrapper"
+                        :style="{left: isShowSecondMenu ? 180 + 'px' : '0'}"
+                >
+                    <transition
+                            name="custom-classes-transition"
+                            enter-active-class="animated fadeIn"
+                    >
+                        <slot name="main-content"/>
+                    </transition>
+                </div>
+            </div>
         </div>
-      </div>
     </div>
-  </div>
 </template>
 
 <script>
-	import http from '../../assets/js/http'
-	import api from '../../api/msgApi'
-	import SideBarNav from '../common/SideBarNav'
-	import SecondNavBar from '../common/SecondNavBar'
-	import TopBarDropMenu from '../common/TopBarDropMenu'
-	import _ from 'lodash'
+    import api from '../../api/msgApi'
+    import SideBarNav from '../common/SideBarNav'
+    import SecondNavBar from '../common/SecondNavBar'
+    import TopBarDropMenu from '../common/TopBarDropMenu'
+    import _ from 'lodash'
 
-	export default {
-  name: 'MainHome',
-  componentName: 'MainHome',
-  components: { SideBarNav, SecondNavBar, TopBarDropMenu },
-  mixins: [http],
-  data () {
-    return {
-      isDropUserMenu: false,
-      unreadMsgCnt: 0,
-      menuList: [],
-      userInfo: {
-        nickname: '',
-        username: '',
-        mobile: '',
-        id: '',
-        head: ''
-      },
-      platformInfo: {
-        all: '',
-        mini: 'BY'
-      },
-      isMiniMenu: false,
-      isShowSecondMenu: false,
-      activeNavIndex: false,
-      activeNavMenuIndex: false,
-      secondMenuData: false,
-      userDropMenus: [
-        { 'name': this.$i18n.t('Avatar'), url: '/admin/account/avatar' },
-        { 'name': this.$i18n.t('ModifyPassword'), url: '/admin/account/password' },
-        { 'name': this.$i18n.t('Logout'), url: '/logout' }
-      ],
-      languages: [
-        {
-            'name': 'English', method: 'changeLanguages', 'value': 'en'
+    export default {
+        name: 'MainHome',
+        componentName: 'MainHome',
+        components: { SideBarNav, SecondNavBar, TopBarDropMenu },
+        data () {
+            return {
+                isDropUserMenu: false,
+                unreadMsgCnt: 0,
+                menuList: [],
+                userInfo: {
+                    nickname: '',
+                    username: '',
+                    mobile: '',
+                    id: '',
+                    head: ''
+                },
+                platformInfo: {
+                    all: '',
+                    mini: 'BY'
+                },
+                isMiniMenu: false,
+                isShowSecondMenu: false,
+                activeNavIndex: false,
+                activeNavMenuIndex: false,
+                secondMenuData: false,
+                userDropMenus: [
+                    { 'name': this.$i18n.t('Avatar'), url: '/admin/account/avatar' },
+                    { 'name': this.$i18n.t('ModifyPassword'), url: '/admin/account/password' },
+                    { 'name': this.$i18n.t('Logout'), url: '/logout' }
+                ],
+                languages: [
+                    {
+                        'name': 'English', method: 'changeLanguages', 'value': 'en'
+                    },
+                    {
+                        'name': '中文简体', method: 'changeLanguages', 'value': 'zh'
+                    }
+                ]
+            }
         },
-        {
-            'name': '中文简体', method: 'changeLanguages', 'value': 'zh'
-        }
-      ]
-    }
-  },
-  computed: {
-    isLoading () {
-      return this.$store.getters.userSessionData.loading
-    },
-    userSessionData () {
-      return this.$store.getters.userSessionData
-    },
-    userAvatarUrl () {
-      return window.tools.getAvatarUrl(this.userInfo.id)
-    },
-    navMaxHeight () {
-      // const navNumber = this.menuList.length;
-      // 顶部导航高度50 侧边导航切换按钮高度30 菜单头40
-      return document.body.clientHeight - 40 * (2 + 2) - 50 - 30
-    }
-  },
-  watch: {
-    userSessionData: function (newValue) {
-      console.debug('用戶數據', newValue)
-      if (newValue.loading === 2) {
-        console.debug('正在请求')
-        return
-      }
-      if (newValue.code === 0) {
-        let data = _.cloneDeep(newValue.data)
-        console.debug('登录成功', data)
-        this.loadMenu(data.menuList)
-        this.loadUserInfo(data.userInfo)
-        this.loadPlatformInfo(data.platformInfo)
-        window.tools.alertClose()
-        this.$router.push('/admin/index')
-      } else if (newValue.msg) {
-        console.debug('登录失败', newValue)
-        window.tools.alertWarn(newValue.msg)
-        setTimeout(() => {
-          this.$router.push('/login')
-        }, 2500)
-      } else {
-        console.debug('登录失败', newValue)
-        setTimeout(() => {
-          this.$router.push('/login')
-        }, 2500)
-      }
-    }
-  },
-  created () {
-    this.getUserData()
-    this.getUnreadMsg()
-  },
-  mounted () {},
-  methods: {
-    jump2AdminIndex() {
-      this.$router.push('/admin/index');
-    },
-  	getUnreadMsg() {
-  	  	api.getUnreadCount({'uid': window.tools.getUID()}, (resp) => {
-  	  		this.unreadMsgCnt = resp
-        }, (resp) => {
-        })
-    },
-  	goMessage() {
-        window.tools.returnTop()
-        window.location.href = this.$router.resolve('/admin/message/index', '#', false).href
-    },
-    changeLanguages (lang) {
-      console.debug('changeLanguages', lang)
-      window.cache.setValue('lang', lang.value, 24*3600)
-      window.location.reload();
-    },
-    getUserData () {
-      window.tools.alertInfo(this.$i18n.t('Loading'))
-      this.$store.dispatch('getUserSessionData')
-    },
-    // 切换迷你侧边导航
-    toggleMiniMenu () {
-      this.isMiniMenu = !this.isMiniMenu
-    },
-    // 切换显示二级菜单
-    toggleSecondMenu () {
-      this.isShowSecondMenu = !this.isShowSecondMenu
-    },
-    // 导航按钮被点击
-    SideBarNavOnClick (index, menu) {
-      this.activeNavMenuIndex = index
-      this.secondMenuData = menu
-      this.isShowSecondMenu = !!menu
-    },
-    // 导航激活点亮
-    SideBarNavOnActive (index) {
-      this.activeNavIndex = index
-    },
-    // 切换全屏
-    toggleFullScreen () {
-      window.screenfull.toggle()
-    },
-    // 刷新iframe
-    iframeRefresh () {
-      console.debug(this.$route.fullPath)
-    },
-    // 退出
-    logout () {
-      console.debug('logout')
-      this.$router.push('logout')
-    },
-    // 链接点击
-    userLinkClick (link) {
-      window.tools.returnTop()
-      // setTimeout(() => { this.$store.dispatch('mainFrameJump', window.tools.getApiUrl(link)) }, 1)
-    },
-    queryChildMenu (data = [], action) {
-      if (typeof action !== 'function') return
-      let ret = []
-      data.forEach(val => {
-        if (action(val)) {
-          val.IsFront = (parseInt(val.url_type) === 1)
-          ret.push(val)
-        }
-      })
-      return ret
-    },
-    loadChildMenu (list = [], data = []) {
-      list = list.map(val => {
-        val.children = this.queryChildMenu(data, (v) => (v.pid == val.id))
-        val.AllowExpand = val.children.length > 0 && val.url === '#'
-        return val
-      })
-      return list
-    },
-    // 菜单数据处理
-    loadMenu (menuList = []) {
-      let newList = this.queryChildMenu(menuList, (v) => (v.pid == 0))
-      newList = this.loadChildMenu(newList, menuList)
-      newList = newList.map(val => {
-        val.children = this.loadChildMenu(val.children, menuList)
-        return val
-      })
-      this.menuList = newList
-    },
-    getApiUrl (url) {
-      return window.tools.getApiUrl(url)
-    },
-    // 加载用户信息
-    loadUserInfo (userInfo) {
-      this.userInfo.nickname = userInfo.nickname
-      this.userInfo.id = userInfo.id
-      this.userInfo.username = userInfo.username
-      this.userInfo.head = userInfo.head
-      this.userInfo.mobile = userInfo.mobile
+        computed: {
+            isLoading () {
+                return this.$store.getters.userSessionData.loading
+            },
+            userSessionData () {
+                return this.$store.getters.userSessionData
+            },
+            userAvatarUrl () {
+                return window.tools.getAvatarUrl(this.userInfo.id)
+            },
+            navMaxHeight () {
+                // const navNumber = this.menuList.length;
+                // 顶部导航高度50 侧边导航切换按钮高度30 菜单头40
+                return document.body.clientHeight - 40 * (2 + 2) - 50 - 30
+            }
+        },
+        watch: {
+            userSessionData: function (newValue) {
+                console.debug('用戶數據', newValue)
+                if (newValue.loading === 2) {
+                    console.debug('正在请求')
+                    return
+                }
+                if (newValue.code === 0) {
+                    let data = _.cloneDeep(newValue.data)
+                    console.debug('登录成功', data)
+                    this.loadMenu(data.menuList)
+                    this.loadUserInfo(data.userInfo)
+                    this.loadPlatformInfo(data.platformInfo)
+                    window.tools.alertClose()
+                    this.$router.push('/admin/index')
+                } else if (newValue.msg) {
+                    console.debug('登录失败', newValue)
+                    window.tools.alertWarn(newValue.msg)
+                    setTimeout(() => {
+                        this.$router.push('/login')
+                    }, 2500)
+                } else {
+                    console.debug('登录失败', newValue)
+                    setTimeout(() => {
+                        this.$router.push('/login')
+                    }, 2500)
+                }
+            }
+        },
+        created () {
 
-      // 替换本地的基础client_id 为 登录用户的clientId
-      window.tools.setClientId(userInfo.client_id)
-      window.tools.setAvatar(userInfo.head)
-    },
-    // 加载平台信息
-    loadPlatformInfo (PlatformInfo) {
-      this.platformInfo.all = PlatformInfo
+        },
+        mounted () {
+          this.getUserData()
+          // this.getUnreadMsg()
+        },
+        methods: {
+            jump2AdminIndex () {
+                this.$router.push('/admin/index')
+            },
+            getUnreadMsg () {
+                api.getUnreadCount({ 'uid': window.tools.getUID() }, (resp) => {
+                    this.unreadMsgCnt = resp
+                }, (resp) => {
+                })
+            },
+            goMessage () {
+                window.tools.returnTop()
+                window.location.href = this.$router.resolve('/admin/message/index', '#', false).href
+            },
+            changeLanguages (lang) {
+                console.debug('changeLanguages', lang)
+                window.cache.setValue('lang', lang.value, 24 * 3600)
+                window.location.reload()
+            },
+            getUserData () {
+                window.tools.alertInfo(this.$i18n.t('Loading'))
+                this.$store.dispatch('getUserSessionData')
+            },
+            // 切换迷你侧边导航
+            toggleMiniMenu () {
+                this.isMiniMenu = !this.isMiniMenu
+            },
+            // 切换显示二级菜单
+            toggleSecondMenu () {
+                this.isShowSecondMenu = !this.isShowSecondMenu
+            },
+            // 导航按钮被点击
+            SideBarNavOnClick (index, menu) {
+                this.activeNavMenuIndex = index
+                this.secondMenuData = menu
+                this.isShowSecondMenu = !!menu
+            },
+            // 导航激活点亮
+            SideBarNavOnActive (index) {
+                this.activeNavIndex = index
+            },
+            // 切换全屏
+            toggleFullScreen () {
+                window.screenfull.toggle()
+            },
+            // 刷新iframe
+            iframeRefresh () {
+                console.debug(this.$route.fullPath)
+            },
+            // 退出
+            logout () {
+                console.debug('logout')
+                this.$router.push('logout')
+            },
+            // 链接点击
+            userLinkClick (link) {
+                window.tools.returnTop()
+                // setTimeout(() => { this.$store.dispatch('mainFrameJump', window.tools.getApiUrl(link)) }, 1)
+            },
+            queryChildMenu (data = [], action) {
+                if (typeof action !== 'function') return
+                let ret = []
+                data.forEach(val => {
+                    if (action(val)) {
+                        val.IsFront = (parseInt(val.url_type) === 1)
+                        ret.push(val)
+                    }
+                })
+                return ret
+            },
+            loadChildMenu (list = [], data = []) {
+                list = list.map(val => {
+                    val.children = this.queryChildMenu(data, (v) => (v.pid == val.id))
+                    val.AllowExpand = val.children.length > 0 && val.url === '#'
+                    return val
+                })
+                return list
+            },
+            // 菜单数据处理
+            loadMenu (menuList = []) {
+                let newList = this.queryChildMenu(menuList, (v) => (v.pid == 0))
+                newList = this.loadChildMenu(newList, menuList)
+                newList = newList.map(val => {
+                    val.children = this.loadChildMenu(val.children, menuList)
+                    return val
+                })
+                this.menuList = newList
+            },
+            getApiUrl (url) {
+                return window.tools.getApiUrl(url)
+            },
+            // 加载用户信息
+            loadUserInfo (userInfo) {
+                this.userInfo.nickname = userInfo.nickname
+                this.userInfo.id = userInfo.id
+                this.userInfo.username = userInfo.username
+                this.userInfo.head = userInfo.head
+                this.userInfo.mobile = userInfo.mobile
+
+                // 替换本地的基础client_id 为 登录用户的clientId
+                window.tools.setClientId(userInfo.client_id)
+                window.tools.setAvatar(userInfo.head)
+            },
+            // 加载平台信息
+            loadPlatformInfo (PlatformInfo) {
+                this.platformInfo.all = PlatformInfo
+            }
+        }
     }
-  }
-}
 </script>
