@@ -84,16 +84,13 @@
 
 <script>
     import goodsApi from '../../api/dtGoodsApi'
-    import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-    import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-    import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
     import ImgUploader2online from '@/components/img-uploader2online.vue'
 
     export default {
         components: {
-            ElForm,
-            ElButtonGroup,
-            ElButton,
             ImgUploader2online
         },
         data () {
