@@ -217,9 +217,9 @@
 
 <script>
     import goodsSkuApi from '../../api/dtGoodsApi'
-    import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-    import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-    import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
     import ImgUploader from '../../components/img-uploader'
 
     export default {
@@ -227,10 +227,7 @@
             id: String
         },
         components: {
-            ImgUploader,
-            ElForm,
-            ElButtonGroup,
-            ElButton
+            ImgUploader
         },
         data() {
             return {
