@@ -506,8 +506,8 @@
 <script>
     import freightApi from '../../api/freightApi'
     import pcaApi from '../../api/pcaApi'
-    import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-    import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
 
     export default {
         components: {
