@@ -197,8 +197,8 @@
 
 <script>
 	import api from '../../api/datatreeApi'
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
 
 	export default {
 		components: {
