@@ -135,9 +135,9 @@
     import goodsSkuApi from '../../api/goodsSkuApi'
     import spCateApi from '../../api/spCateApi'
     import goodsApi from '../../api/goodsApi'
-    import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-    import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-    import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
     import ImgUploader from '../../components/img-uploader'
 
     export default {
@@ -145,10 +145,7 @@
             id: String
         },
         components: {
-            ImgUploader,
-            ElForm,
-            ElButtonGroup,
-            ElButton
+            ImgUploader
         },
         data () {
             return {
