@@ -283,15 +283,13 @@
 
 <script>
 	import spPropApi from '../../api/spPropApi'
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
 
 	export default {
 		components: {
-			ElForm,
-			ElButtonGroup,
-			ElButton
+
 		},
 		data() {
 			return {
