@@ -171,9 +171,9 @@
 
 <script>
 	import spPropValueApi from '../../api/spPropValueApi'
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
 
 	export default {
 		props: {
@@ -181,9 +181,7 @@
         },
 
 		components: {
-			ElForm,
-			ElButtonGroup,
-			ElButton
+
 		},
 		data() {
 			return {
