@@ -175,9 +175,9 @@
 <script>
     import spShopApi from '../../api/spShopApi'
     import goodsApi from '../../api/goodsApi'
-    import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-    import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-    import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
 
     export default {
         props: {
@@ -189,9 +189,6 @@
             }
         },
         components: {
-            ElForm,
-            ElButtonGroup,
-            ElButton
         },
         data () {
             return {
