@@ -126,16 +126,14 @@
 
 <script>
 	import spCateApi from '../../api/spCateApi'
-	import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-	import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-	import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
 	import spPropApi from "../../api/spPropApi"
 
 	export default {
 		components: {
-			ElForm,
-			ElButtonGroup,
-			ElButton
+
 		},
 		data() {
 			return {
