@@ -430,17 +430,14 @@
   import api from '../../api/videoApi'
   import fileApi from '../../api/fileApi'
   import cateApi from '../../api/videoCateApi'
-  import ElButton from '../../../node_modules/element-ui/packages/button/src/button.vue'
-  import ElButtonGroup from '../../../node_modules/element-ui/packages/button/src/button-group.vue'
-  import ElForm from '../../../node_modules/element-ui/packages/form/src/form.vue'
+
+
+
   import ImgUploader from '@/components/img-uploader2online.vue'
 
   export default {
     components: {
-      ElForm,
-      ElButtonGroup,
-      ElButton,
-      ImgUploader
+        ImgUploader
     },
     data () {
       return {
