@@ -1,16 +1,14 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
 import user from './modules/user'
 import datatree from './modules/datatree'
-
-Vue.use(Vuex)
+import {createStore} from 'vuex'
 
 const debug = process.env.NODE_ENV !== 'production'
 
-export default new Vuex.Store({
+const store = createStore({
   modules: {
     user,
     datatree
   },
   strict: debug
 })
+export default  store
